requirements.txt
asgiref==3.2.3
Django==3.0.2
django-crispy-forms==1.8.1
django-extensions==2.2.6
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
numpy==1.18.1
opencv-python==4.1.2.30
Pillow==7.0.0
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
